refactor: 상품 카테고리 및 정렬 필터, 데이터 페칭 훅으로 분리
